Number field default cast from int to float
It allows displaying default float numbers.

+++ src/Form/Field/Number.php
@@ -10,7 +10,7 @@ class Number extends Text
 
     public function render()
     {
-        $this->default((int) $this->default);
+        $this->default((float) $this->default);
 
         $this->script = <<<EOT
